+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 1.0.2-alpha.3
+* [#47] - Fix EditText.applyFilterAllowedDigits behavior
+* [#48] - Add new extensions and classes for EditText:
+    - Added ```EditText.applyFilters(bindFunction: ArrayList<InputFilter>.() -> Unit)```
+    - Added ```inline fun EditText.listenTextChanges(crossinline function: (text: CharSequence?) -> Unit)```
+    - Added ```EditText.makeSecure()```
+    - Added ```EditText.toggleSecure()```
+    - Created class ```AllowedCharactersInputFilter```
+    - Created class ```OnlyDigitsInputFilter```
+    - Created class ```OnlyLettersOrDigitsInputFilter```
+    - Added fabric methods to create such input filters as:
+     ```AllowedCharactersInputFilter```,```OnlyDigitsInputFilter```,
+     ```OnlyLettersOrDigitsInputFilter```, ```IntentFilter.AllCaps```,
+     ```IntentFilter.LengthFilter```
+    - Created abstract class ```SimpleTextWatcher``` 
 # 1.0.1-alpha.2 (Jul 11, 2018)
 * [#33] - Fixed typo at SupportedFeaturesExtensions
 * Update kotlin and gradle version